+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 * Updating to ISO v4 and MAGIC Discovery v1 record schemas
+* Reorganise static components separately from base components
 
 ### Removed